Update signatures of generated functions in tests/exec to new lowering procedure
When lowering FuncOps with memref arguments, newer versions of MLIR
generate one argument per field of the structure describing the memref
instead of generating a single argument per memref that points to the
structure describing the memref. This patch updates the signatures and
invocations of generated functions in tests/exec to the new lowering
procedure.