Analyses and transformations for normalizing (affine) arithmetic expressions.
The collection of 'monomials' is the most central of the analyses:
it identifies the variables and their respective coefficients (which
generally consist of constants and parameters) in an expression.
The results of the collection of 'monomials' analysis are the most
useful for, e.g., normalization if the input expression has been
suitably transformed, in particular by distributing multiplication
over addition/subtraction.